+++ Leaderboards/Actions/GetPlayerScores.cs
@@ -8,13 +8,13 @@ namespace ConstructServices.Leaderboards.Actions
     {
         public static GetScoreResponse GetPlayerScores(
             this LeaderboardService service,
-            string playerIdentifier,
+            string playerID,
             PaginationOptions paginationOptions)
         {
             const string path = "/getplayerscores.json";
             var formData = new Dictionary<string, string>
             {
-                { "playerID", playerIdentifier }
+                { "playerID", playerID }
             };
             return Task.Run(() => Request.ExecuteLeaderboardRequest<GetScoreResponse>(
                 path,