+++ Leaderboards/Actions/TeamsRemovePlayer.cs
@@ -10,18 +10,18 @@ namespace ConstructServices.Leaderboards.Actions
         public static BaseResponse RemovePlayerFromTeam(
             this LeaderboardService service,
             string strTeamID,
-            string playerIdentifier)
+            string playerID)
         {
             if (string.IsNullOrWhiteSpace(strTeamID))
                 return new BaseResponse("No Team ID was provided.", false);
             if (!Guid.TryParse(strTeamID, out var teamID))
                 return new BaseResponse("Score Team was not a valid GUID.", false);
-            return RemovePlayerFromTeam(service, teamID, playerIdentifier);
+            return RemovePlayerFromTeam(service, teamID, playerID);
         }
         public static BaseResponse RemovePlayerFromTeam(
             this LeaderboardService service,
             Guid teamID,
-            string playerIdentifier)
+            string playerID)
         {
             const string path = "/removeplayerfromteam.json";
 
@@ -31,7 +31,7 @@ namespace ConstructServices.Leaderboards.Actions
                 new Dictionary<string, string>
                 {
                     { "teamID", teamID.ToString() },
-                    { "playerID", playerIdentifier }
+                    { "playerID", playerID }
                 }
             )).Result;
         }