+++ Leaderboards/RequestPerspective.cs
@@ -1,15 +1,18 @@
-﻿namespace ConstructServices.Leaderboards
+﻿using System;
+
+namespace ConstructServices.Leaderboards
 {
     public class RequestPerspective
     {
         internal string RequesterIP { get; set; }
-        internal string RequesterPlayerID { get; set; }
+        internal Guid? RequesterPlayerID { get; set; }
         
         public RequestPerspective(string requesterIP)
         {
             RequesterIP = requesterIP;
+            RequesterPlayerID = null;
         }
-        public RequestPerspective(string requesterIP, string requesterPlayerID)
+        public RequestPerspective(string requesterIP, Guid? requesterPlayerID)
         {
             RequesterIP = requesterIP;
             RequesterPlayerID = requesterPlayerID;