+++ CHANGELOG.md
@@ -1,4 +1,16 @@
 ﻿# Changelog
 
+## 1.0.7.0 - 1st July 2025
+
+<ul>
+	<li>New authentication service that allows you to get and register players, returning their corresponding player ID's.</li>
+	<li>Various minor bug fixes and improvements.</li>
+</ul>
+
+#### BREAKING CHANGES:
+<ul>
+	<li>Leaderboard player identifiers have been renamed to `playerID` parameters which expect a Guid value.</li>
+</ul>
+
 ## 1.0.6.0 - 29th April 2025
 Initial beta release